Retrieve Shared Folders for a Specific User
This change will expose a new endpoint that was recently created in the API to allow a caller to request the list of Shared Folders for a specific user.  The current functionality only allows for the Shared Folders of the currently authenticated user to be returned.  This new call can only be made by a member of the Super User group within the Account.